+++ src/geometries/geometries.jl
@@ -13,7 +13,7 @@ import .User:
     Annulus, Annuli,
     Triangle, Mesh,
     Ring, BendyCylinder,
-    load_mesh, SWCFile, SWCNode, read_swc, read_swc_as_spheres,
+    load_mesh, SWCFile, SWCNode, read_swc, read_swc_raw,
     random_positions_radii, nvolumes,
     write_geometry, read_geometry_json, read_geometry
 end